requirements.txt
alembic==1.4.3
Flask==1.1.2
Flask-Assets==2.0
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Jinja2==2.11.2
MarkupSafe==1.1.1
python-dateutil==2.8.1
SQLAlchemy==1.3.20
webassets==2.0
Werkzeug==1.0.1
